+++ packages/react/src/components/RichText/CoreEditor/Extensions/Mention/types.tsx
@@ -16,7 +16,7 @@ export interface MentionItemComponentProps {
 }
 
 export type MentionedUser = {
-  id: number
+  id: string | number
   label: string
   image_url?: string
   href?: string